adding first unit tests for code builder based gen
