--- test/goldens/test_b.dart.golden
@@ -1,79 +1,0 @@
-part of typewriter.testing.test_b;
-
-const jsonCodec = const JsonCodec();
-
-class PersonDecoder extends Converter<Object, Person> {
-  const PersonDecoder();
-
-  @override
-  Person convert(Object raw) {
-    var input = raw as Map<String, dynamic>;
-    var output = new Person();
-    output.items = input["items"].map((x) => x).toList();
-    output.age = input["age"];
-    output.coins = input["coins"];
-    output.dog = new Dog(input["dog"]);
-    output.cat = const CatDecoder().convert(input["cat"]);
-    return output;
-  }
-}
-
-class PersonEncoder extends Converter<Person, Object> {
-  const PersonEncoder();
-
-  @override
-  Object convert(Person input) {
-    var output = <String, dynamic>{};
-    output["items"] = input.items.map((x) => x).toList();
-    output["age"] = input.age;
-    output["coins"] = input.coins;
-    output["dog"] = input.dog.encode();
-    output["cat"] = const CatEncoder().convert(input.cat);
-    return output;
-  }
-}
-
-class PersonCodec extends Codec<Person, Object> {
-  const PersonCodec();
-
-  @override
-  Converter<Object, Person> get decoder => const PersonDecoder();
-
-  @override
-  Converter<Person, Object> get encoder => const PersonEncoder();
-}
-
-class CatDecoder extends Converter<Object, Cat> {
-  const CatDecoder();
-
-  @override
-  Cat convert(Object raw) {
-    var input = raw as Map<String, dynamic>;
-    var output = new Cat();
-    output.name = input["name"];
-    output.birthday = DateTime.parse(input["birth_day"]);
-    return output;
-  }
-}
-
-class CatEncoder extends Converter<Cat, Object> {
-  const CatEncoder();
-
-  @override
-  Object convert(Cat input) {
-    var output = <String, dynamic>{};
-    output["name"] = input.name;
-    output["birth_day"] = input.birthday.toIso8601String();
-    return output;
-  }
-}
-
-class CatCodec extends Codec<Cat, Object> {
-  const CatCodec();
-
-  @override
-  Converter<Object, Cat> get decoder => const CatDecoder();
-
-  @override
-  Converter<Cat, Object> get encoder => const CatEncoder();
-}
